reuse code between tests
